rpm: Store OAuth2 encryption key in a data file rather than in code
SuiteCRM generates an OAuth2 encryption key and stores it by modifying
the code file containing the ApiConfig class.  This is unbelievably
dumb.

Question to SuiteCRM developers: when you find yourself writing code
that reads in another PHP code file, performs a string substitution,
and writes it back out, do you not ever think "there must be a better
way to do this"?  Like, for example, using an actual configuration
file.  That way, you can upgrade the code without destroying your
configuration, or vice versa.  It's magic!

It's also really not that hard.  The rest of the world has been doing
it the right way since at least the 1960s, so you've only got half a
century of coding practices to catch up on.  Get here when you can.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>